+++ templates/blog/blog.js
@@ -1,5 +1,8 @@
 import { styleCaption } from '../../scripts/scripts.js';
-import { div } from '../../scripts/dom-helpers.js';
+import { 
+  a, div, h3, p 
+} from '../../scripts/dom-helpers.js';
+import { createOptimizedPicture } from '../../scripts/lib-franklin.js';
 
 export default function buildAutoBlocks() {
   // add social share block
@@ -44,5 +47,41 @@ export default function buildAutoBlocks() {
         ),
       ),
     ),
+  );
+}
+
+function summariseDescription(description) {
+  let result = description;
+  if (result.length > 75) {
+    result = result.substring(0, 75);
+    const lastSpaceIndex = result.lastIndexOf(' ');
+    if (lastSpaceIndex !== -1) {
+      result = result.substring(0, lastSpaceIndex);
+    }
+  }
+  return `${result}…'`;
+}
+
+export function renderBlogCard(item) {
+  const itemImage = item.thumbnail && item.thumbnail !== '0' ? item.thumbnail : item.image;
+  const buttonText = item.cardC2A && item.cardC2A !== '0' ? item.cardC2A : 'Read More';
+
+  return (
+    div({ class: 'blog-card' },
+      div({ class: 'blog-card-thumb' },
+        a({ href: item.path },
+          createOptimizedPicture(itemImage, item.title, 'lazy', [{ width: '800' }]),
+        ),
+      ),
+      div({ class: 'blog-card-caption' },
+        h3(
+          a({ href: item.path }, item.title),
+        ),
+        p({ class: 'blog-card-description' }, summariseDescription(item.description)),
+        p({ class: 'button-container' },
+          a({ href: item.path, 'aria-label': buttonText, class: 'button primary' }, buttonText),
+        ),
+      ),
+    )
   );
 }
